Move install tests to `CommandBuilder`
