+++ tests/command_builder/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+from statue.command_builder import CommandBuilder
+
+
+@pytest.fixture
+def mock_get_package(mocker):
+    return mocker.patch.object(CommandBuilder, "_get_package")
