+++ tests/command_builder/test_command_builder_install.py
@@ -1,27 +1,29 @@
 import sys
 
-from statue.command import Command
+from statue.command_builder import CommandBuilder
 from statue.verbosity import SILENT
 from tests.constants import COMMAND1, COMMAND_HELP_STRING1
 
 
-def test_command_wont_install_if_already_installed(mock_get_package, mock_subprocess):
+def test_command_builder_wont_install_if_already_installed(
+    mock_get_package, mock_subprocess
+):
     version = "6.2.1"
     mock_get_package.return_value.version = version
-    command = Command(name=COMMAND1, help=COMMAND_HELP_STRING1)
+    command_builder = CommandBuilder(name=COMMAND1, help=COMMAND_HELP_STRING1)
 
-    command.install()
+    command_builder.install()
 
     mock_subprocess.assert_not_called()
 
 
-def test_command_install_if_not_already_installed(
+def test_command_builder_install_if_not_already_installed(
     mock_get_package, mock_subprocess, environ
 ):
     mock_get_package.return_value = None
-    command = Command(name=COMMAND1, help=COMMAND_HELP_STRING1)
+    command_builder = CommandBuilder(name=COMMAND1, help=COMMAND_HELP_STRING1)
 
-    command.install()
+    command_builder.install()
 
     mock_subprocess.assert_called_once_with(
         [sys.executable, "-m", "pip", "install", COMMAND1],
@@ -31,11 +33,11 @@ def test_command_install_if_not_already_installed(
     )
 
 
-def test_command_install_silently(mock_get_package, mock_subprocess, environ):
+def test_command_builder_install_silently(mock_get_package, mock_subprocess, environ):
     mock_get_package.return_value = None
-    command = Command(name=COMMAND1, help=COMMAND_HELP_STRING1)
+    command_builder = CommandBuilder(name=COMMAND1, help=COMMAND_HELP_STRING1)
 
-    command.install(verbosity=SILENT)
+    command_builder.install(verbosity=SILENT)
 
     mock_subprocess.assert_called_once_with(
         [sys.executable, "-m", "pip", "install", COMMAND1],
@@ -45,12 +47,16 @@ def test_command_install_silently(mock_get_package, mock_subprocess, environ):
     )
 
 
-def test_command_install_specified_version(mock_get_package, mock_subprocess, environ):
+def test_command_builder_install_specified_version(
+    mock_get_package, mock_subprocess, environ
+):
     version = "4.2.1"
     mock_get_package.return_value = None
-    command = Command(name=COMMAND1, help=COMMAND_HELP_STRING1, version=version)
+    command_builder = CommandBuilder(
+        name=COMMAND1, help=COMMAND_HELP_STRING1, version=version
+    )
 
-    command.install()
+    command_builder.install()
 
     mock_subprocess.assert_called_once_with(
         [sys.executable, "-m", "pip", "install", f"{COMMAND1}=={version}"],
